introduced a tiny node module that gets the project version from pom.xml and is shared by gulp and karma
